stm-ai-service/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

class Settings:
    OPENROUTER_API_KEY = os.getenv("OPENROUTER_API_KEY")
    OPENROUTER_URL = os.getenv("OPENROUTER_URL")
    DB_HOST = os.getenv("DB_HOST", "localhost")
    DB_USER = os.getenv("DB_USER", "root")
    DB_PASSWORD = os.getenv("DB_PASSWORD", "P@ssw0rd")
    DB_NAME = os.getenv("DB_NAME", "smarttask_db")

settings = Settings()