stm-ai-service/requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
mysql-connector-python
httpx[http2]
edge-tts
pydub
vosk